Add file verification entry into shard (#39)
Add FileVerificationEntry into shard. If the most significant bit of FileDataSequenceHeader::file_flags is set, there are equal number of FileVerificationEntrys following FileDataSequenceEntrys.

When cleaning a file completes, one keyed-hash is computed for each range of chunk hashes in a cas object for verification, with key being the cas object hash.