Drop xet_error. (#134)
xet_error was a source drop of thiserror with integrated logging and the ability to put in a breakpoint to note exception propagation. However, the benefits of maintaining a source drop with changes, at least on client side, are less clear in the current ecosystem and with upgrades to the thiserror package (e.g. better integrated stacktracing).

This PR removes xet_error and reverts everything back to the original thiserror implementation.