Drop on shard_file_manager issues warning in debug mode instead of error. (#128)
Currently, the drop in the shard file manager is present as a programming guard to make sure the user has called finish and written all the in-memory lookup tables to disk. However, this isn't a valid check on runtime shutdown, so this switches it to happen only in debug mode so a user hitting CTRL-C won't see scary errors.