Profiler  (#143)
This PR adds in an easy way to profile xet-core operations. Internally,
the code uses pprof-rs, which links in the GPerfTools suite.

To use, compile with --features=profiling; with maturin, install with
`maturin develop --release --features=profiling`


When running, it saves the profile to profilers/<timestamp>, writing
both a gperf protobuf file and a flamegraph svg file. The protobuf file
can be read using the pprof tool:

```bash
go install github.com/google/pprof@latest
# Replace this with the correct path to the hf_xet library file.
~/go/bin/pprof  ./venv/lib/python3.10/site-packages/hf_xet/hf_xet.abi3.so  profiles/<timestamp>/pprof.pb
```

On OSX, I like using qcachegrind as the interactive visualizer of the
results, which requires the profile data to be exported to the callgrind
format:

```bash
brew install qcachegrind     # If needed
~/go/bin/pprof  -callgrind ./venv/lib/python3.10/site-packages/hf_xet/hf_xet.abi3.so  profiles/<timestamp>/pprof.pb
qcachegrind <reported callgrind output file> 
```