Caching system for shard file manager (#159)
This PR enables a caching system for the shard file manager by: 
- shard file manager is associated with a single directory. All
registered shards must be in that directory.
- Within-process caching is optionally done by directory.   
- The shard file manager can optionally reference a secondary manager
for chunk and file lookup. In our case, the main cache directory is
cached and global, but each session directory has its own shard file
manager where lookup is attempted first, falling back to the global
cache directory manager if that fails.