Rename data processing class names to reflect functionality  (#176)
Renamed the following to more clearly reflect the current functionality:
- PointerFileTranslator is split into FileUploadSession and
FileDownloader
- Cleaner -> SingleFileCleaner

The mental model for a call to upload_files is to define a session, where a session may translate and upload multiple files and should fully complete, including uploading all data, before returning.  This renaming makes this it clearer where the functionality of different parts of this model live.