Clean up global dedup query API  (#180)
Changed the trait ShardDedupProber to return a PathBuf instead of just
the shard hash for the global dedup, as the shard hash was reformed to
get the path but with unclear location. Also, as this is used only in
the cas_client, moved the trait there.

Furthermore, corrected an issue with the local testing client and added
test coverage for that.

This fixes the issue where global dedup shards were downloaded to the
cache but attempted to be registered by the session shard manager.