Xorb Format Upgrade (#182)
Upgrade to support uncompressed lengths stored in the xorb footer; other
rearrangements as needed.

In detail,

- Rename `CasObjectInfo` to `CasObjectInfoV0`
- Add `CasObjectInfoV1` that implements the new footer format.
- Change `CasObject` to contain the new footer. A new object will be
serialized to use the new footer. When deserializing a xorb, it will
check the footer version and dispatch the corresponding footer
deserialization logic.
- Augments xorb validation from async stream to support three cases:
xorb with V1 footer, xorb with V0 footer, and xorb with no footer. For
the last two cases, the footer is disregarded and a new V1 footer is
generated using information gained from chunk list validation and
returned to the caller.

---------

Co-authored-by: Assaf Vayner <assaf@huggingface.co>
Co-authored-by: seanses <di@huggingface.co>