Restrict memory allocation size on Xorb deserialization (#197)
Limit the initial variable length array allocation size and increment
the array size as needed. This is to prevent allocating giant array
based on the size declared in an uploaded xorb. This will pair up with
the caller limiting the total read size of a reader to effectively limit total
memory usage.