opt-test profile for builds with debug assertions for regular use (#210)
This PR ups the optimization level for the opt-test profile to 3 to
better enable this mode to be used for non-release but production
testing. In this mode, debug_assertions are enabled, which can catch
errors early on or ensure integrity checks are enabled. Otherwise, it
behaves just like release mode.