Macro for declaring configurable constants.  (#211)
Currently, we use a lot of const expressions in which a corresponding
environment variable can be used to override the default value. This PR
introduces a macro to standardize this usage.

Example:
```
use utils::configurable_constants;

configurable_constants! {

    /// The target shard size; shards.
    ref MDB_SHARD_TARGET_SIZE: u64 = 64 * 1024 * 1024;

    /// Minimum shard size; shards are aggregated until they are at least this.
    ref MDB_SHARD_MIN_TARGET_SIZE: u64 = 64 * 1024 * 1024;

    /// The global dedup chunk modulus; a chunk is considered global dedup
    /// eligible if the hash modulus this value is zero.  Can only be changed for testing.
    ref MDB_SHARD_GLOBAL_DEDUP_CHUNK_MODULUS : u64 = release_fixed(1024);
}
```

In this example, on initialization, the code would attempt to read from
XET_ MDB_SHARD_TARGET_SIZE, using the value given above if that doesn't
work.

For MDB_SHARD_GLOBAL_DEDUP_CHUNK_MODULUS, which is denoted by
release_fixed, it would attempt to load the value from the corresponding
environment variable only when debug_assertions are enabled, allowing
that value to be changed in tests but not in release mode. Similarly,
fixed(...) denotes that a value is never changed.

Additionally, the macro `test_set_global!(VAR, VALUE);` can be called at
the beginning of a test to ensure that the global referenced by that
value is set to a specific value for that test.