Env to write reconstruction terms in parallel (#221)
When investigating slowness between hf_xet download and hf_transfer, @Hugoch
noticed that one source of slowness was due to the fact that although
terms were downloaded concurrently, they were being written sequentially
to disk. In
[hf_transfer](https://github.com/huggingface/hf_transfer/blob/main/src/lib.rs#L334-L361),
each chunk downloaded is written to the output file directly.

This PR updates the way the cas_client writes reconstructed terms to the
output file to behave similarly to hf_transfer:
* Refactors the reconstruction path to take a `OutputProvider` enum
instead of a `&Box<dyn Write + Send>`.
* `OutputProvider` enum can be called to provide a `Write`
implementation for a particular location.
* New method: `RemoteClient::reconstruct_file_to_writer_parallel` to
write terms to the output file in parallel.
* Env variable: `HF_XET_RECONSTRUCT_WRITE_SEQUENTIALLY` that, when set
by the user, will enable writing terms to the output sequentially
instead of in parallel.