Update hf_xet to 1.0.0 (#222)
Now that we've released the integration with huggingface_hub and users
are expected to use hf_xet in production, we should bump up the version
of hf_xet to reflect this.

This does imply that we will not be breaking the interface to hf_xet
(i.e. the pyo3 interface), however, I think that that is something to
adhere to as breaking that interface will break current huggingface_hub
versions.